requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
chinesecalendar==1.0.6
click==6.7
cycler==0.10.0
decorator==4.3.0
Distance==0.1.3
enum34==1.1.6
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
hyperdash==0.15.2
idna==2.6
jieba==0.39
Keras==2.1.6
kiwisolver==1.0.1
lightgbm==2.1.0
Markdown==2.6.11
matplotlib==2.2.2
networkx==2.1
nltk==3.3
numpy==1.14.2
pandas==0.22.0
Pillow==5.1.0
plac==0.9.6
pprint==0.1
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.2
python-slugify==1.2.5
pytz==2018.4
PyYAML==3.12
requests==2.18.4
sacrebleu==1.2.7
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
sumeval==0.1.6
tensorboard==1.8.0
tensorboardX==1.4
tensorflow-gpu==1.5.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
torch==0.4.1
torchvision==0.2.1
tqdm==4.23.2
typing==3.6.4
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
gensim